+++ assembly/src/ast/parsers/mod.rs
@@ -9,6 +9,7 @@ use core::{fmt::Display, ops::RangeBounds};
 
 mod adv_ops;
 mod debug;
+mod emit;
 mod field_ops;
 mod io_ops;
 mod stack_ops;