+++ miden/tests/integration/operations/decorators/mod.rs
@@ -1,2 +1,3 @@
 mod advice;
 mod asmop;
+mod event;